+++ playground/src/countPerUser/CountPerUser.tsx
@@ -6,15 +6,17 @@ interface CountPerUserProps {
 export function CountPerUser({ username }: CountPerUserProps) {
   const [count, set_count, allValues] =
     useStateTogetherWithPerUserValues<number>('countPerUser', 0, {
-      persistDisconnectedUserData: true,
-      keyOverride: username,
+      // keepValues: true,
       // resetOnConnect: true,
       // resetOnDisconnect: true,
-      overwriteSessionValue: true
+      // overwriteSessionValue: true
+      //
+      keyOverride: username
     })
   return (
     <>
-      <strong>{username}: </strong>
+      {/* <strong>{username}: </strong> */}
+      <span>My value:</span>
       <button
         onClick={() => set_count((p) => p + 1)}
         className="bg-neutral-500 py-1 px-3 rounded"